GEN_AI/requirements.txt
# Core FastAPI Backend
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.7.1
pydantic-settings==2.5.2
python-dotenv==1.0.0

# Job Storage & Caching
redis==5.1.1

# Retry Logic
tenacity==9.0.0

# AI & Computer Vision
torch==2.5.1
torchvision==0.20.1
numpy==2.1.3
opencv-python==4.12.0.88
pillow==10.2.0

# Audio & Speech Processing
moviepy==1.0.3
openai-whisper==20231117
ffmpeg-python==0.2.0

# Web Interface (Optional - for local testing)
streamlit==1.52.2

# Testing (Optional - for development)
pytest==8.3.4
pytest-asyncio==0.24.0
